+++ examples/stm32/nucleo_l432kc/hello_world/Src/main.c
@@ -4,9 +4,31 @@
  * @copyright Copyright (c) 2020-2026 Tropic Square s.r.o.
  *
  * @license For the license see LICENSE.md in the root directory of this source tree.
+ */
+
+// This project is based on SPI_FullDuplex_ComPolling example by MCD Application Team
+// and modified by Tropic Square to run Libtropic functional tests, see original
+// file header attached below.
+
+/**
+ ******************************************************************************
+ * @file    SPI/SPI_FullDuplex_ComPolling/Src/main.c
+ * @author  MCD Application Team
+ * @brief   This sample code shows how to use STM32L4xx SPI HAL API to transmit
+ *          and receive a data buffer with a communication process based on
+ *          Polling transfer.
+ *          The communication is done using 2 Boards.
+ ******************************************************************************
+ * @attention
  *
- * This example project is based on the SPI/SPI_FullDuplex_ComPolling example from STM32 example
- * library which was created by the MCD Application Team.
+ * Copyright (c) 2017 STMicroelectronics.
+ * All rights reserved.
+ *
+ * This software is licensed under terms that can be found in the LICENSE file
+ * in the root directory of this software component.
+ * If no LICENSE file comes with this software, it is provided AS-IS.
+ *
+ ******************************************************************************
  */
 
 /* Includes ------------------------------------------------------------------*/
@@ -17,7 +39,7 @@
 
 #include "libtropic.h"
 #include "libtropic_mbedtls_v4.h"
-#include "libtropic_port_stm32_nucleo_l432kc.h"
+#include "libtropic_port_stm32l4xx.h"
 #include "psa/crypto.h"
 
 /** @addtogroup STM32L4xx_HAL_Examples
@@ -37,11 +59,11 @@
 
 /* Choose pairing keypair for slot 0. */
 #if LT_USE_SH0_ENG_SAMPLE
-#define LT_EX_SH0_PRIV sh0priv_eng_sample
-#define LT_EX_SH0_PUB sh0pub_eng_sample
+#define LT_EX_SH0_PRIV lt_sh0priv_eng_sample
+#define LT_EX_SH0_PUB lt_sh0pub_eng_sample
 #elif LT_USE_SH0_PROD0
-#define LT_EX_SH0_PRIV sh0priv_prod0
-#define LT_EX_SH0_PUB sh0pub_prod0
+#define LT_EX_SH0_PRIV lt_sh0priv_prod0
+#define LT_EX_SH0_PUB lt_sh0pub_prod0
 #endif
 
 /* Private macro -------------------------------------------------------------*/
@@ -226,7 +248,7 @@ int main(void)
 
         The device structure has to be zero initialized!
         STM32 HAL depends on zero init values. */
-    lt_dev_stm32_nucleo_l432kc_t device = {0};
+    lt_dev_stm32l4xx_t device = {0};
 
     device.spi_instance = LT_SPI_INSTANCE;
     device.baudrate_prescaler = SPI_BAUDRATEPRESCALER_16;