+++ hal/stm32/stm32l4xx/libtropic_port_stm32l4xx.c
@@ -1,12 +1,12 @@
 /**
- * @file libtropic_port_stm32_nucleo_l432kc.c
+ * @file libtropic_port_stm32l4xx.c
  * @copyright Copyright (c) 2020-2026 Tropic Square s.r.o.
- * @brief Port for STM32 L432KC using native SPI HAL (and GPIO HAL for chip select).
+ * @brief Port for STM32L4xx series using native SPI HAL (and GPIO HAL for chip select).
  *
  * @license For the license see LICENSE.md in the root directory of this source tree.
  */
 
-#include "libtropic_port_stm32_nucleo_l432kc.h"
+#include "libtropic_port_stm32l4xx.h"
 
 #include <stdarg.h>
 #include <stdint.h>
@@ -20,11 +20,11 @@
 #include "libtropic_secure_memzero.h"
 #include "stm32l4xx_hal.h"
 
-#define LT_STM32_L432KC_GPIO_OUTPUT_CHECK_ATTEMPTS 10
+#define LT_STM32L4XX_GPIO_OUTPUT_CHECK_ATTEMPTS 10
 
 lt_ret_t lt_port_random_bytes(lt_l2_state_t *s2, void *buff, size_t count)
 {
-    lt_dev_stm32_nucleo_l432kc_t *device = (lt_dev_stm32_nucleo_l432kc_t *)(s2->device);
+    lt_dev_stm32l4xx_t *device = (lt_dev_stm32l4xx_t *)(s2->device);
     size_t bytes_left = count;
     uint8_t *buff_ptr = buff;
     int ret;
@@ -50,10 +50,10 @@ lt_ret_t lt_port_random_bytes(lt_l2_state_t *s2, void *buff, size_t count)
 
 lt_ret_t lt_port_spi_csn_low(lt_l2_state_t *s2)
 {
-    lt_dev_stm32_nucleo_l432kc_t *device = (lt_dev_stm32_nucleo_l432kc_t *)(s2->device);
+    lt_dev_stm32l4xx_t *device = (lt_dev_stm32l4xx_t *)(s2->device);
 
     HAL_GPIO_WritePin(device->spi_cs_gpio_bank, device->spi_cs_gpio_pin, GPIO_PIN_RESET);
-    for (uint8_t read_attempts = 0; read_attempts < LT_STM32_L432KC_GPIO_OUTPUT_CHECK_ATTEMPTS;
+    for (uint8_t read_attempts = 0; read_attempts < LT_STM32L4XX_GPIO_OUTPUT_CHECK_ATTEMPTS;
          read_attempts++) {
         if (!HAL_GPIO_ReadPin(device->spi_cs_gpio_bank, device->spi_cs_gpio_pin)) {
             return LT_OK;
@@ -66,10 +66,10 @@ lt_ret_t lt_port_spi_csn_low(lt_l2_state_t *s2)
 
 lt_ret_t lt_port_spi_csn_high(lt_l2_state_t *s2)
 {
-    lt_dev_stm32_nucleo_l432kc_t *device = (lt_dev_stm32_nucleo_l432kc_t *)(s2->device);
+    lt_dev_stm32l4xx_t *device = (lt_dev_stm32l4xx_t *)(s2->device);
 
     HAL_GPIO_WritePin(device->spi_cs_gpio_bank, device->spi_cs_gpio_pin, GPIO_PIN_SET);
-    for (uint8_t read_attempts = 0; read_attempts < LT_STM32_L432KC_GPIO_OUTPUT_CHECK_ATTEMPTS;
+    for (uint8_t read_attempts = 0; read_attempts < LT_STM32L4XX_GPIO_OUTPUT_CHECK_ATTEMPTS;
          read_attempts++) {
         if (HAL_GPIO_ReadPin(device->spi_cs_gpio_bank, device->spi_cs_gpio_pin)) {
             return LT_OK;
@@ -82,7 +82,7 @@ lt_ret_t lt_port_spi_csn_high(lt_l2_state_t *s2)
 
 lt_ret_t lt_port_init(lt_l2_state_t *s2)
 {
-    lt_dev_stm32_nucleo_l432kc_t *device = (lt_dev_stm32_nucleo_l432kc_t *)(s2->device);
+    lt_dev_stm32l4xx_t *device = (lt_dev_stm32l4xx_t *)(s2->device);
     int ret;
 
     // Set the SPI parameters.
@@ -120,12 +120,21 @@ lt_ret_t lt_port_init(lt_l2_state_t *s2)
     GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_MEDIUM;
     HAL_GPIO_Init(device->spi_cs_gpio_bank, &GPIO_InitStruct);
 
+#if LT_USE_INT_PIN
+    // GPIO for INT pin.
+    GPIO_InitStruct.Pin = device->int_gpio_pin;
+    GPIO_InitStruct.Mode = GPIO_MODE_INPUT;
+    GPIO_InitStruct.Pull = GPIO_NOPULL;
+    GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
+    HAL_GPIO_Init(device->int_gpio_bank, &GPIO_InitStruct);
+#endif
+
     return LT_OK;
 }
 
 lt_ret_t lt_port_deinit(lt_l2_state_t *s2)
 {
-    lt_dev_stm32_nucleo_l432kc_t *device = (lt_dev_stm32_nucleo_l432kc_t *)(s2->device);
+    lt_dev_stm32l4xx_t *device = (lt_dev_stm32l4xx_t *)(s2->device);
     int ret;
 
     ret = HAL_SPI_DeInit(&device->spi_handle);
@@ -140,7 +149,7 @@ lt_ret_t lt_port_deinit(lt_l2_state_t *s2)
 lt_ret_t lt_port_spi_transfer(lt_l2_state_t *s2, uint8_t offset, uint16_t tx_data_length,
                               uint32_t timeout_ms)
 {
-    lt_dev_stm32_nucleo_l432kc_t *device = (lt_dev_stm32_nucleo_l432kc_t *)(s2->device);
+    lt_dev_stm32l4xx_t *device = (lt_dev_stm32l4xx_t *)(s2->device);
 
     if (offset + tx_data_length > TR01_L1_LEN_MAX) {
         LT_LOG_ERROR("Invalid data length!");
@@ -164,6 +173,25 @@ lt_ret_t lt_port_delay(lt_l2_state_t *s2, uint32_t ms)
 
     return LT_OK;
 }
+
+#if LT_USE_INT_PIN
+lt_ret_t lt_port_delay_on_int(lt_l2_state_t *s2, uint32_t ms)
+{
+    lt_dev_stm32l4xx_t *device = (lt_dev_stm32l4xx_t *)(s2->device);
+    uint32_t time_initial = HAL_GetTick();
+    uint32_t time_actual;
+
+    while ((HAL_GPIO_ReadPin(device->int_gpio_bank, device->int_gpio_pin) == 0)) {
+        time_actual = HAL_GetTick();
+        if ((time_actual - time_initial) > ms) {
+            return LT_L1_INT_TIMEOUT;
+        }
+        // HAL_Delay(ms);
+    }
+
+    return LT_OK;
+}
+#endif
 
 int lt_port_log(const char *format, ...)
 {